+++ .github/workflows/03_code_review.yml
@@ -18,7 +18,7 @@ on:
   workflow_dispatch:
 
 env:
-  PROJECT_KEY: pagopa_pagopa-node-cfg-job-manager
+  PROJECT_KEY: pagopa_pagopa-node-cfg-sync
 
 
 permissions:
