+++ .github/workflows/05h_deploy_with_github_runner.yml
@@ -13,7 +13,7 @@ on:
         type: string
 
 env:
-  APP_NAME: pagopa-node-cfg-job-manager
+  APP_NAME: pagopa-node-cfg-sync
 
 
 permissions:
