+++ src/main/java/it/gov/pagopa/node/cfg_sync/config/MappingsConfiguration.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.microservice.config;
+package it.gov.pagopa.node.cfg_sync.config;
 
 
 import org.modelmapper.ModelMapper;
