+++ src/main/java/it/gov/pagopa/node/cfg_sync/exception/AppError.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.microservice.exception;
+package it.gov.pagopa.node.cfg_sync.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
