+++ src/main/java/it/gov/pagopa/node/cfg_sync/exception/ErrorHandler.java
@@ -1,8 +1,9 @@
-package it.gov.pagopa.microservice.exception;
+package it.gov.pagopa.node.cfg_sync.exception;
 
-import it.gov.pagopa.microservice.model.ProblemJson;
+import it.gov.pagopa.node.cfg_sync.model.ProblemJson;
 import java.util.ArrayList;
 import java.util.List;
+
 import lombok.extern.slf4j.Slf4j;
 import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.beans.TypeMismatchException;
