+++ src/main/resources/application.properties
@@ -1,5 +1,6 @@
 # Info
 info.application.artifactId=@project.artifactId@
+info.application.name=@project.name@
 info.application.version=@project.version@
 info.application.description=@project.description@
 info.properties.environment=${ENV:azure}
@@ -14,7 +15,6 @@ management.health.readinessState.enabled=true
 springdoc.writer-with-order-by-keys=true
 springdoc.writer-with-default-pretty-printer=true
 # Server
-# TODO: set your base path
 server.servlet.context-path=/
 server.port=8080
 # Logging
