+++ src/test/java/it/gov/pagopa/node/cfg_sync/ApplicationTest.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.microservice;
+package it.gov.pagopa.node.cfg_sync;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
